Pick up the correct scope for rewrites
This fixes the 'optimized-float-expression' test, which now passes.
The problem before was that a condition was incorrectly added based
on the existence in the (wrong) scope of a bound argument 'x',
but then we (correctly) did not know how to evaluate x.

The condition check corresponds to declarations like:

     x + x is 2 * x

In that case, a runtime check is made on a call like [A + B] to check
that A = B.