Clarify the scopes and lookup for subexpressions during type inference
This makes it possible to pass the type analysis in optimized-self-comma-list.

The scenario is something like:

    X, Y is self
    "Hello", "World", 42

When we try to build the type inference for the list, the type and scope
of X have to be new for every sub-element. For example, we start binding
[X] to ["Hello"] and [Y] to ["World", 42]. That binding should not interfere
with the subsequent binding of [X] to ["World"], even if it has a
different type.