A more efficient way to inject evaluated trees into the rcalls map
The method from the previous commit was copying all the expressions,
which was unnecessarily expensive, and would have polluted the rcalls
map for subexpressions with all calls from the parent types.

The approach here is to only copy one RewriteCalls at a time right
after evaluating it in the subexpressions's type system.