Make sure we lookup arguments in the parameters scope
When we generate the body of a function like [foo X is decl],
we generate one scope for local parameter bindings (e.g. [X])
and one enclosed scope for whatever is declared in [decl].

Therefore, when looking [X] up, we will not find it in the [decl]
scope but only in the enclosing scope.